完成 人事档案，事件记录 模块

+++ FCARDIO.Protocol.Fingerprint/Fingerprint/Person/WritePersonBase.cs
@@ -78,7 +78,7 @@ namespace FCARDIO.Protocol.Fingerprint.Person
             }
             buf.Clear();
 
-            buf.WriteInt(iLen);
+            buf.WriteByte(iLen);
 
             for (int i = 0; i < iLen; i++)
             {
